remove data race and add readme

+++ philo_bonus/check_args.c
@@ -20,14 +20,14 @@ int	check_args(int ac, char **av, t_args *args)
 		args->time_to_die = ft_atoi(av[2]);
 		args->time_to_eat = ft_atoi(av[3]);
 		args->time_to_sleep = ft_atoi(av[4]);
-		if (args->nb_philo < 0 || args->time_to_die < 0 \
-			|| args->time_to_eat < 0 || args->time_to_sleep < 0)
+		if ((int)(args->nb_philo) < 0 || (int)(args->time_to_die) < 0 \
+			|| (int)(args->time_to_eat) < 0 || (int)(args->time_to_sleep) < 0)
 			return (0);
 		if (ac == 6)
 			args->nbr_time_to_eat = ft_atoi(av[5]);
 		else
 			args->nbr_time_to_eat = (size_t)(-1);
-		if (args->nbr_time_to_eat < 0 && args->nbr_time_to_eat != (size_t)(-1))
+		if ((int)(args->nbr_time_to_eat) < 0 && (int)(args->nbr_time_to_eat) != (-1))
 			return (0);
 		return (1);
 	}
